pyproject.toml
[project]
name = "ai-recommend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marcelo Ortiz de Santana", email = "tentativafc@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "dependency-injector>=4.46.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "jupyterlab>=4.5.4",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "opentelemetry-distro>=0.60b1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "pandas>=3.0.0",
    "pgvector>=0.3.6",
    "plotly>=6.5.2",
    "polars>=1.38.1",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
]

[project.scripts]
ai-recommend = "ai_recommend:main"


[dependency-groups]
dev = [
    "jupyter-ruff>=0.3.1",
    "pandas-stubs>=3.0.0.260204",
    "ruff>=0.15.1",
]

[tool.ruff]
# Target Python 3.14
target-version = "py314"
line-length = 100
indent-width = 4

# Enable all linter rules by default
lint.select = ["E", "F", "W", "I", "N", "D", "UP", "YTT", "ANN", "ASYNC", "S", "B", "A", "C4", "T10", "EM", "ISC", "G", "INP", "PIE", "T20", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "ARG", "PTH", "TD", "FIX", "ERA", "PL", "TRY", "PERF", "FURB", "LOG", "RUF"]

# Ignore specific rules
lint.ignore = [
    "E501",    # Line too long (handled by formatter)
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "S101",    # Use of assert (allowed in tests)
    "PLR0913", # Too many arguments to function definition
    "PLR2004", # Magic value used in comparison
]

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*" = ["S101", "S102", "ANN"]

# Formatter configuration
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

# Docstring style: Google
[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["overload"]

# Isort configuration (import sorting)
[tool.ruff.lint.isort]
known-first-party = ["ai_recommend"]
known-third-party = ["dependency_injector", "sqlalchemy", "pgvector", "loguru", "pydantic"]
lines-after-imports = 2
force-single-line = false
combine-as-imports = true
case-sensitive = true
force-wrap-aliases = false

# Naming conventions
[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

# Pylint rules
[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "str", "bytes"]
max-args = 7
max-branches = 13
max-returns = 6
max-statements = 50


